requirement.txt
# requirements.txt
openai
requests
python-dotenv
selenium
webdriver-manager